refactor: extract FilteredClassInfo-related classes
